Implements a configurable "one-liner" option
For some cases, the added newlines & indents of short objects and arrays
makes the output unwieldy. This commit adds a new option 
`inlineCharacterLimit`, which when set to a positive Number will 
return stringified object or array that, with padding/indents/newlines
stripped would have a length less than `inlineCharacterLimit`.

As background, if I use stringify-object to print this argument list:

`foo(obj1, obj2)` 

And `obj1` is `{a: 1, b: 2}` and `obj2` is `[1,2,3]`, and I try to 
print a description of the invocation back to the user like this:

`'foo(' + stringifyObject(obj1) + ', ' + stringifyObject(obj2) + ')'`

Then what could be nice and terse as: 

`'foo({a: 1, b: 2}, [1,2,3])'`

But will instead print something containing a lot of new lines, making 
it hard to see the invocation being described.

I'm running into this a lot developing
https://github.com/testdouble/testdouble.js to have better output,
because using terse object and array params to pass to test doubles is 
tremendously common, and I want to give users as best output as I can.